% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\docType{data}
\name{Especes}
\alias{Especes}
\title{Classification phylogenetique}
\format{Un \code{data.frame} avec 11 colonnes :
\describe{
  \item{Classe}{Classe de l'espèce}
  \item{Ordre}{Ordre de l'espèce}
  \item{Sous-ordre}{Sous-ordre de l'espèce}
  \item{Famille}{Famille de l'espèce}
  \item{Genre}{Genre de l'espèce}
  \item{Nom vernaculaire}{Nom vernaculaire de l'espèce}
  \item{Nom scientifique}{Nom scientifique de l'espèce}
  \item{Taille}{Taille de l'espèce}
  \item{Description}{Description de l'espèce}
  \item{Sexe}{Sexe de l'espèce pour les cas de dimorphisme sexuel}
  \item{Img}{Nom de l'image de l'espèce}
}}
\usage{
Especes
}
\description{
Classification phylogénétique, descriptions et lien vers les photos des espèces d'insectes et d'oiseaux
}
\keyword{datasets}
