% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Extraire_Classe}
\alias{F_Extraire_Classe}
\title{Extraction de la classe}
\usage{
F_Extraire_Classe(df_Row)
}
\arguments{
\item{df_Row}{\code{data.frame} avec une colonne \code{Classe}}
}
\value{
\code{character} avec la classe extraite
}
\description{
Permet d'extraire la classe d'une espèce du \code{data.frame} \code{df_Row}
}
\examples{
RandomRow = F_Random_Row(Especes)

F_Extraire_Classe(RandomRow)

}
