% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Extraire_Description}
\alias{F_Extraire_Description}
\title{Extraction de la description}
\usage{
F_Extraire_Description(df_Row)
}
\arguments{
\item{df_Row}{\code{data.frame} avec une colonne \code{Description}}
}
\value{
\code{character} avec la description extraite
}
\description{
Permet d'extraire la description d'une espèce du \code{data.frame} \code{df_Row}
}
\examples{
RandomRow = F_Random_Row(Especes)

F_Extraire_Description(RandomRow)

}
