% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Extraire_Famille}
\alias{F_Extraire_Famille}
\title{Extraction de la famille}
\usage{
F_Extraire_Famille(df_Row)
}
\arguments{
\item{df_Row}{\code{data.frame} avec une colonne \code{Famille}}
}
\value{
\code{character} avec la famille extraite
}
\description{
Permet d'extraire la famille d'une espèce du \code{data.frame} \code{df_Row}
}
\examples{
RandomRow = F_Random_Row(Especes)

F_Extraire_Famille(RandomRow)

}
