% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Extraire_Genre}
\alias{F_Extraire_Genre}
\title{Extraction du genre}
\usage{
F_Extraire_Genre(df_Row)
}
\arguments{
\item{df_Row}{\code{data.frame} avec une colonne \code{Genre}}
}
\value{
\code{character} avec le genre extrait
}
\description{
Permet d'extraire le genre d'une espèce du \code{data.frame} \code{df_Row}
}
\examples{
RandomRow = F_Random_Row(Especes)

F_Extraire_Genre(RandomRow)

}
