% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Extraire_Img}
\alias{F_Extraire_Img}
\title{Extraction de l'image}
\usage{
F_Extraire_Img(df_Row)
}
\arguments{
\item{df_Row}{\code{data.frame} avec une colonne \code{Img}}
}
\value{
\code{character} avec le nom de l'image extraite
}
\description{
Permet d'extraire l'image d'une espèce du \code{data.frame} \code{df_Row}
}
\examples{
RandomRow = F_Random_Row(Especes)

F_Extraire_Img(RandomRow)

}
