% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Extraire_Ordre}
\alias{F_Extraire_Ordre}
\title{Extraction de l'ordre}
\usage{
F_Extraire_Ordre(df_Row)
}
\arguments{
\item{df_Row}{\code{data.frame} avec une colonne \code{Ordre}}
}
\value{
\code{character} avec l'ordre extrait
}
\description{
Permet d'extraire l'ordre d'une espèce du \code{data.frame} \code{df_Row}
}
\examples{
RandomRow = F_Random_Row(Especes)

F_Extraire_Ordre(RandomRow)

}
