% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Extraire_Sexe}
\alias{F_Extraire_Sexe}
\title{Extraction du sexe}
\usage{
F_Extraire_Sexe(df_Row)
}
\arguments{
\item{df_Row}{\code{data.frame} avec une colonne \code{Sexe}}
}
\value{
\code{character} avec le sexe extrait
}
\description{
Permet d'extraire le sexe d'une espèce du \code{data.frame} \code{df_Row}
}
\examples{
RandomRow = F_Random_Row(Especes)

F_Extraire_Sexe(RandomRow)

}
