% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Extraire_SousFamille}
\alias{F_Extraire_SousFamille}
\title{Extraction de la sous-famille}
\usage{
F_Extraire_SousFamille(df_Row)
}
\arguments{
\item{df_Row}{\code{data.frame} avec une colonne \code{Sous-famille}}
}
\value{
\code{character} avec la sous-famille extraite
}
\description{
Permet d'extraire la sous-famille d'une espèce du \code{data.frame} \code{df_Row}
}
\examples{
RandomRow = F_Random_Row(Especes)

F_Extraire_SousFamille(RandomRow)

}
