% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Extraire_SousOrdre}
\alias{F_Extraire_SousOrdre}
\title{Extraction du sous-ordre}
\usage{
F_Extraire_SousOrdre(df_Row)
}
\arguments{
\item{df_Row}{\code{data.frame} avec une colonne \code{Sous-ordre}}
}
\value{
\code{character} avec le sous-ordre extrait
}
\description{
Permet d'extraire le sous-ordre d'une espèce du \code{data.frame} \code{df_Row}
}
\examples{
RandomRow = F_Random_Row(Especes)

F_Extraire_SousOrdre(RandomRow)

}
