% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Extraire_Sp}
\alias{F_Extraire_Sp}
\title{Extraction de l'espece}
\usage{
F_Extraire_Sp(df_Row)
}
\arguments{
\item{df_Row}{\code{data.frame} avec une colonne \code{Nom vernaculaire}}
}
\value{
\code{character} avec le nom vernaculaire extrait
}
\description{
Permet d'extraire le nom vernaculaire d'une espèce du \code{data.frame} \code{df_Row}
}
\examples{
RandomRow = F_Random_Row(Especes)

F_Extraire_Sp(RandomRow)

}
