% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Extraire_Sp_sci}
\alias{F_Extraire_Sp_sci}
\title{Extraction du nom scientifique}
\usage{
F_Extraire_Sp_sci(df_Row)
}
\arguments{
\item{df_Row}{\code{data.frame} avec une colonne \code{Nom scientifique}}
}
\value{
\code{character} avec le nom scientifique extrait
}
\description{
Permet d'extraire le nom scientifique d'une espèce du \code{data.frame} \code{df_Row}
}
\examples{
RandomRow = F_Random_Row(Especes)

F_Extraire_Sp_sci(RandomRow)

}
