% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Extraire_Taille}
\alias{F_Extraire_Taille}
\title{Extraction de la taille}
\usage{
F_Extraire_Taille(df_Row)
}
\arguments{
\item{df_Row}{\code{data.frame} avec une colonne \code{Taille}}
}
\value{
\code{character} avec la taille extraite
}
\description{
Permet d'extraire la taille d'une espèce du \code{data.frame} \code{df_Row}
}
\examples{
RandomRow = F_Random_Row(Especes)

F_Extraire_Taille(RandomRow)

}
