% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Filtre_Par_Classe}
\alias{F_Filtre_Par_Classe}
\title{Filtre par classe}
\usage{
F_Filtre_Par_Classe(df_Espece, classe)
}
\arguments{
\item{df_Espece}{\code{data.frame} avec une colonne \code{Classe}}
}
\value{
\code{data.frame} \code{df_Especes} filtré selon la classe souhaitée
}
\description{
Permet de filtrer le \code{data.frame} \code{df_Especes} selon la classe
}
\examples{
F_Filtre_Par_Classe(Especes, classe = "Aves")
F_Filtre_Par_Classe(Especes, classe = "Insectes")

}
