% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Filtre_Par_Famille}
\alias{F_Filtre_Par_Famille}
\title{Filtre par famille}
\usage{
F_Filtre_Par_Famille(df_Espece, famille)
}
\arguments{
\item{df_Espece}{\code{data.frame} avec une colonne \code{Famille}}
}
\value{
\code{data.frame} \code{df_Especes} filtré selon la famille souhaitée
}
\description{
Permet de filtrer le \code{data.frame} \code{df_Especes} selon la famille
}
\examples{
F_Filtre_Par_Famille(Especes, famille = "Pieridae")

library(dplyr)
Especes \%>\%
 F_Filtre_Par_Ordre(ordre = "Odonates") \%>\%
 F_Filtre_Par_Famille(famille = "Coenagrionidae")

}
