% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Filtre_Par_Genre}
\alias{F_Filtre_Par_Genre}
\title{Filtre par genre}
\usage{
F_Filtre_Par_Genre(df_Espece, genre)
}
\arguments{
\item{df_Espece}{\code{data.frame} avec une colonne \code{Genre}}
}
\value{
\code{data.frame} \code{df_Especes} filtré selon le genre souhaité
}
\description{
Permet de filtrer le \code{data.frame} \code{df_Especes} selon le genre
}
\examples{
F_Filtre_Par_Genre(Especes, genre = "Polyommatus")

library(dplyr)
Especes \%>\%
 F_Filtre_Par_Famille(famille = "Coenagrionidae") \%>\%
 F_Filtre_Par_Genre(genre = "Pyrrhosoma")


}
