% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Filtre_Par_Ordre}
\alias{F_Filtre_Par_Ordre}
\title{Filtre par ordre}
\usage{
F_Filtre_Par_Ordre(df_Espece, ordre)
}
\arguments{
\item{df_Espece}{\code{data.frame} avec une colonne \code{Ordre}}
}
\value{
\code{data.frame} \code{df_Especes} filtré selon l'ordre souhaité
}
\description{
Permet de filtrer le \code{data.frame} \code{df_Especes} selon l'ordre
}
\examples{
F_Filtre_Par_Ordre(Especes, ordre = "Odonates")

library(dplyr)
Especes \%>\%
 F_Filtre_Par_Classe(classe = "Insectes") \%>\%
 F_Filtre_Par_Ordre(ordre = "Odonates")

}
