% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Filtre_Par_SousFamille}
\alias{F_Filtre_Par_SousFamille}
\title{Filtre par sous-famille}
\usage{
F_Filtre_Par_SousFamille(df_Espece, sousfamille)
}
\arguments{
\item{df_Espece}{\code{data.frame} avec une colonne \code{Sous-famille}}
}
\value{
\code{data.frame} \code{df_Especes} filtré selon la sous-famille souhaitée
}
\description{
Permet de filtrer le \code{data.frame} \code{df_Especes} selon la sous-famille
}
\examples{
library(dplyr)
Especes \%>\%
 F_Filtre_Par_Famille(famille = "Anatidae") \%>\%
 F_Liste_SousFamille()

}
