% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Filtre_Par_SousOrdre}
\alias{F_Filtre_Par_SousOrdre}
\title{Filtre par sous-ordre}
\usage{
F_Filtre_Par_SousOrdre(df_Espece, sousordre)
}
\arguments{
\item{df_Espece}{\code{data.frame} avec une colonne \code{Sous-ordre}}
}
\value{
\code{data.frame} \code{df_Especes} filtré selon le sous-ordre souhaité
}
\description{
Permet de filtrer le \code{data.frame} \code{df_Especes} selon le sous-ordre
}
\examples{
library(dplyr)
Vec_SousOrdre = Especes \%>\%
 F_Filtre_Par_Ordre(ordre = "Odonates") \%>\%
 F_Liste_SousOrdre()

Especes \%>\%
 F_Filtre_Par_SousOrdre(sousordre = Vec_SousOrdre[1])

}
