% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Liste_Classe}
\alias{F_Liste_Classe}
\title{Liste des classes}
\usage{
F_Liste_Classe(df_Espece)
}
\arguments{
\item{df_Espece}{\code{data.frame} avec une colonne \code{Classe}}
}
\value{
Un vecteur avec la liste des classes du \code{data.frame} \code{df_Especes}
}
\description{
Permet d'obtenir la liste des classes du \code{data.frame} \code{df_Especes}
}
\examples{
F_Liste_Classe(Especes)

}
