% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Liste_Famille}
\alias{F_Liste_Famille}
\title{Liste des familles}
\usage{
F_Liste_Famille(df_Espece)
}
\arguments{
\item{df_Espece}{\code{data.frame} avec une colonne \code{Famille}}
}
\value{
Un vecteur avec la liste des familles du \code{data.frame} \code{df_Especes}
}
\description{
Permet d'obtenir la liste des familles du \code{data.frame} \code{df_Especes}
}
\examples{
F_Liste_Famille(Especes)

library(dplyr)
Especes \%>\%
 F_Filtre_Par_Ordre(ordre = "Odonates") \%>\%
 F_Liste_Famille()

}
