% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Liste_Genre}
\alias{F_Liste_Genre}
\title{Liste des genres}
\usage{
F_Liste_Genre(df_Espece)
}
\arguments{
\item{df_Espece}{\code{data.frame} avec une colonne \code{Genre}}
}
\value{
Un vecteur avec la liste des genre du \code{data.frame} \code{df_Especes}
}
\description{
Permet d'obtenir la liste des genres du \code{data.frame} \code{df_Especes}
}
\examples{
F_Liste_Genre(Especes)

library(dplyr)
Especes \%>\%
 F_Filtre_Par_Famille(famille = "Coenagrionidae") \%>\%
 F_Liste_Genre()

}
