% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Liste_Ordre}
\alias{F_Liste_Ordre}
\title{Liste des ordres}
\usage{
F_Liste_Ordre(df_Espece)
}
\arguments{
\item{df_Espece}{\code{data.frame} avec une colonne \code{Ordre}}
}
\value{
Un vecteur avec la liste des ordres du \code{data.frame} \code{df_Especes}
}
\description{
Permet d'obtenir la liste des ordres du \code{data.frame} \code{df_Especes}
}
\examples{
F_Liste_Ordre(Especes)

library(dplyr)
Especes \%>\%
 F_Filtre_Par_Classe(classe = "Insectes") \%>\%
 F_Liste_Ordre()

}
