% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Liste_Sexe}
\alias{F_Liste_Sexe}
\title{Liste des sexes}
\usage{
F_Liste_Sexe(df_Espece)
}
\arguments{
\item{df_Espece}{\code{data.frame} avec une colonne \code{Sexe}}
}
\value{
Un vecteur avec la liste des sexes du \code{data.frame} \code{df_Especes}
}
\description{
Permet d'obtenir la liste des sexes du \code{data.frame} \code{df_Especes}
}
\examples{
F_Liste_Sexe(Especes)

library(dplyr)
Especes \%>\%
 F_Filtre_Par_Genre(genre = "Eristalis") \%>\%
 F_Liste_Sexe()

}
