% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Liste_SousFamille}
\alias{F_Liste_SousFamille}
\title{Liste des sous-familles}
\usage{
F_Liste_SousFamille(df_Espece)
}
\arguments{
\item{df_Espece}{\code{data.frame} avec une colonne \code{Sous-famille}}
}
\value{
Un vecteur avec la liste des sous-famille du \code{data.frame} \code{df_Especes}
}
\description{
Permet d'obtenir la liste des sous-familles du \code{data.frame} \code{df_Especes}
}
\examples{
F_Liste_SousFamille(Especes)

library(dplyr)
Especes \%>\%
 F_Filtre_Par_Famille(famille = "Anatidae") \%>\%
 F_Liste_SousFamille()

}
