% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Liste_SousOrdre}
\alias{F_Liste_SousOrdre}
\title{Liste des sous-ordres}
\usage{
F_Liste_SousOrdre(df_Espece)
}
\arguments{
\item{df_Espece}{\code{data.frame} avec une colonne \code{Sous-ordre}}
}
\value{
Un vecteur avec la liste des sous-ordres du \code{data.frame} \code{df_Especes}
}
\description{
Permet d'obtenir la liste des sous-ordres du \code{data.frame} \code{df_Especes}
}
\examples{
F_Liste_SousOrdre(Especes)

library(dplyr)
Especes \%>\%
 F_Filtre_Par_Ordre(ordre = "Odonates") \%>\%
 F_Liste_SousOrdre()

}
