% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Liste_Sp}
\alias{F_Liste_Sp}
\title{Liste des especes}
\usage{
F_Liste_Sp(df_Espece)
}
\arguments{
\item{df_Espece}{\code{data.frame} avec une colonne \code{Nom vernaculaire}}
}
\value{
Un vecteur avec la liste des noms vernaculaires des espèces du \code{data.frame} \code{df_Especes}
}
\description{
Permet d'obtenir la liste des noms vernaculaires des espèces du \code{data.frame} \code{df_Especes}
}
\examples{
F_Liste_Sp(Especes)

library(dplyr)
Especes \%>\%
 F_Filtre_Par_Famille(famille = "Coenagrionidae") \%>\%
 F_Liste_Sp()

}
