% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Liste_Sp_sci}
\alias{F_Liste_Sp_sci}
\title{Liste des noms scientifiques}
\usage{
F_Liste_Sp_sci(df_Espece)
}
\arguments{
\item{df_Espece}{\code{data.frame} avec une colonne \code{Nom scientifique}}
}
\value{
Un vecteur avec la liste des noms scientifique des espèces du \code{data.frame} \code{df_Especes}
}
\description{
Permet d'obtenir la liste des noms scientifiques des espèces du \code{data.frame} \code{df_Especes}
}
\examples{
F_Liste_Sp_sci(Especes)

library(dplyr)
Especes \%>\%
 F_Filtre_Par_Famille(famille = "Coenagrionidae") \%>\%
 F_Liste_Sp_sci()

}
