% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Photo_Espece}
\alias{F_Photo_Espece}
\title{Photo de l'espece}
\usage{
F_Photo_Espece(df_Row)
}
\arguments{
\item{df_Row}{\code{data.frame} avec une colonne \code{Img}}
}
\value{
Un plot avec la photo de l'espece
}
\description{
Permet d'afficher la photo d'une espèce du \code{data.frame} \code{df_Row}
}
\examples{
RandomRow = F_Random_Row(Especes)
F_Photo_Espece(RandomRow)

library(dplyr)
Especes \%>\%
 F_Filtre_Par_Famille("Coenagrionidae") \%>\%
 F_Random_Row() \%>\%
 F_Photo_Espece()

}
