% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Random_Classe}
\alias{F_Random_Classe}
\title{Tirage aleatoire d'une classe}
\usage{
F_Random_Classe(df_Especes)
}
\arguments{
\item{df_Espece}{\code{data.frame} avec une colonne \code{Classe}}
}
\value{
\code{character} avec la classe tirée aléatoirement
}
\description{
Permet de tirer aléatoirement une classe du \code{data.frame} \code{df_Especes}
}
\examples{
F_Random_Classe(Especes)

}
