% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Random_Famille}
\alias{F_Random_Famille}
\title{Tirage aleatoire d'une famille}
\usage{
F_Random_Famille(df_Especes, ordre = NULL, sousordre = NULL)
}
\arguments{
\item{ordre}{Ordre dans lequel nous voulons tirer aléatoirement.
Argument facultatif (par défaut \code{ordre = NULL})}

\item{sousordre}{Sous-ordre dans lequel nous voulons tirer aléatoirement.
Argument facultatif (par défaut \code{sousordre = NULL})}

\item{df_Espece}{\code{data.frame} avec une colonne \code{Famille}}
}
\value{
\code{character} avec la famille tirée aléatoirement
}
\description{
Permet de tirer aléatoirement une famille du \code{data.frame} \code{df_Especes}
}
\examples{
F_Random_Famille(Especes)

F_Random_Famille(Especes, ordre = "Odonates")

F_Random_Famille(Especes, sousordre = "Apocrites")

}
