% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Random_Genre}
\alias{F_Random_Genre}
\title{Tirage aleatoire d'un genre}
\usage{
F_Random_Genre(df_Especes, famille = NULL)
}
\arguments{
\item{famille}{Famille dans laquelle nous voulons tirer aléatoirement.
Argument facultatif (par défaut \code{famille = NULL})}

\item{df_Espece}{\code{data.frame} avec une colonne \code{Genre}}
}
\value{
\code{character} avec le genre tiré aléatoirement
}
\description{
Permet de tirer aléatoirement un genre du \code{data.frame} \code{df_Especes}
}
\examples{
F_Random_Genre(Especes)

F_Random_Genre(Especes, famille = "Mantidae")

}
