% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Random_Ordre}
\alias{F_Random_Ordre}
\title{Tirage aleatoire d'un ordre}
\usage{
F_Random_Ordre(df_Especes, classe = NULL)
}
\arguments{
\item{classe}{Classe dans laquelle nous voulons tirer aléatoirement.
Argument facultatif (par défaut \code{classe = NULL})}

\item{df_Espece}{\code{data.frame} avec une colonne \code{Ordre}}
}
\value{
\code{character} avec l'ordre tiré aléatoirement
}
\description{
Permet de tirer aléatoirement un ordre du \code{data.frame} \code{df_Especes}
}
\examples{
F_Random_Ordre(Especes)

F_Random_Ordre(Especes, classe = "Insectes")

}
