% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Random_Sp}
\alias{F_Random_Sp}
\title{Tirage aleatoire d'une espece}
\usage{
F_Random_Sp(df_Especes, genre = NULL)
}
\arguments{
\item{genre}{Genre dans lequel nous voulons tirer aléatoirement.
Argument facultatif (par défaut \code{genre = NULL})}

\item{df_Espece}{\code{data.frame} avec une colonne \code{Nom vernaculaire}}
}
\value{
\code{character} avec le nom vernaculaire tiré aléatoirement
}
\description{
Permet de tirer aléatoirement le nom vernaculaire d'une espèce du \code{data.frame} \code{df_Especes}
}
\examples{
F_Random_Sp(Especes)

F_Random_Sp(Especes, genre = "Cerambyx")

}
