% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Random_Sp_sci}
\alias{F_Random_Sp_sci}
\title{Tirage aleatoire du nom scientifique d'une espece}
\usage{
F_Random_Sp_sci(df_Especes, genre = NULL)
}
\arguments{
\item{genre}{Genre dans lequel nous voulons tirer aléatoirement.
Argument facultatif (par défaut \code{genre = NULL})}

\item{df_Espece}{\code{data.frame} avec une colonne \code{Nom scientifique}}
}
\value{
\code{character} avec le nom scientifique tiré aléatoirement
}
\description{
Permet de tirer aléatoirement le nom scientifique d'une espèce du \code{data.frame} \code{df_Especes}
}
\examples{
F_Random_Sp_sci(Especes)

F_Random_Sp_sci(Especes, genre = "Cygnus")

}
