% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{F_Tirage_Especes}
\alias{F_Tirage_Especes}
\title{Tirage aleatoire de 3 especes proches}
\usage{
F_Tirage_Especes(df_Espece)
}
\arguments{
\item{df_Espece}{\code{data.frame} sur lequel on veut réaliser le tirage aléatoire des espèces}
}
\value{
Un vecteur de\code{character} de longueur 3 avec les 3 espèces
}
\description{
Permet de tirer aléatoirement 3 espèces du \code{data.frame} \code{df_Espece}
}
\examples{
F_Tirage_Especes(Especes)

library(dplyr)
Especes \%>\%
 F_Filtre_Par_Famille(famille = "Anatidae") \%>\%
 F_Tirage_Especes()

}
